Move extractCalendarChange to bot src/util
